+++ packages/components/ellipsize-text/src/ellipsize-text.ts
@@ -21,7 +21,7 @@ declare global {
  */
 export class EllipsizeText extends ScopedElementsMixin(LitElement) {
   /** @internal */
-  static get scopedElements(): ScopedElementsMap {
+  static override get scopedElements(): ScopedElementsMap {
     return {
       'sl-tooltip': Tooltip
     };