+++ packages/components/tag/src/tag-list.ts
@@ -49,7 +49,7 @@ export class TagList extends ScopedElementsMixin(LitElement) {
   }
 
   /** @internal */
-  static get scopedElements(): ScopedElementsMap {
+  static override get scopedElements(): ScopedElementsMap {
     return {
       'sl-tag': Tag
     };