+++ packages/components/toggle-button/src/toggle-button.ts
@@ -47,7 +47,7 @@ export type ToggleButtonSize = 'sm' | 'md' | 'lg';
  */
 export class ToggleButton extends ForwardAriaMixin(ScopedElementsMixin(LitElement)) {
   /** @internal */
-  static get scopedElements(): ScopedElementsMap {
+  static override get scopedElements(): ScopedElementsMap {
     return {
       'sl-icon': Icon,
       'sl-tooltip': Tooltip