+++ packages/components/virtual-list/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @sl-design-system/virtual-list
 
+## 0.1.0
+
+### Minor Changes
+
+- [#3409](https://github.com/sl-design-system/components/pull/3409) [`7d96c3a`](https://github.com/sl-design-system/components/commit/7d96c3aebdc8922f0b031f2ea84aa04c12db2c59) - Migrate to TanStack Virtual for improved performance and reliability
+
+  The virtual-list component has been refactored to use TanStack Virtual instead of the custom virtualization implementation. This provides better performance, smoother scrolling, and more reliable item measurement.
+  - Migrated to TanStack Virtual for core virtualization logic
+  - Added `scrollMargin` property to control scroll offset when scrolling items into view
+  - Added `requestLayout()` method to trigger re-measurement of item sizes
+  - Updated `VirtualListItemRenderer` type to support both Element and TemplateResult return types
+
+### Patch Changes
+
+- Updated dependencies [[`b19dbe7`](https://github.com/sl-design-system/components/commit/b19dbe7d6bffbf3f7e1373f4bcc5693b4352c3ba)]:
+  - @sl-design-system/shared@0.12.2
+
 ## 0.0.5
 
 ### Patch Changes